dt is now in ms, stars can now update variably

+++ Stardroid/app/src/main/java/com/seakernel/stardroid/GameFragment.java
@@ -248,7 +248,7 @@ public class GameFragment extends Fragment implements GLSurfaceView.Renderer {
     @Override
     public void onDrawFrame(GL10 gl) {
         Profiler profiler = Profiler.getInstance();
-        long dt = profiler.trackFrame(mStardroidEngine);
+        float dt = profiler.trackFrame(mStardroidEngine);
 
         StardroidModel model = StardroidModel.getInstance(); // TODO: Definitely change the model to hold stars, and everything else
         GLES20.glClear(GLES20.GL_COLOR_BUFFER_BIT);
